Allow object cache to be forcefully included (#391)
* Add support for rsync-ing mu-plugins and adding object-cache.php during testing

* Ignore object cache

* Lock to 85b3aff18f181fd2c7a08dfd639fa4688fe1179a to debug

* lock to the https://github.com/alleyinteractive/mantle-ci/pull/10 branch

* Testing without

* Restore debug code

* Disable fast-fail

* Disable object cache

* Disable the branch lock in

+++ src/mantle/testing/concerns/trait-rsync-installation.php
@@ -184,13 +184,7 @@ trait Rsync_Installation {
 		}
 
 		// Check if Memcached is installed.
-		if ( ! class_exists( \Memcached::class ) ) {
-			// Allow the object cache to be forcefully required.
-			if ( Utils::env( 'MANTLE_REQUIRE_OBJECT_CACHE', false ) ) {
-				Utils::error( 'Memcached is not installed. Cannot install object cache. Exiting...' );
-				exit( 1 );
-			}
-
+		if ( ! class_exists( \Memcached::class ) && ! Utils::env( 'MANTLE_REQUIRE_OBJECT_CACHE', false ) ) {
 			Utils::error( 'Memcached is not installed. Cannot install object cache. Skipping...' );
 
 			return $this;
